Fix compile error
/home/anatol/sources/dreamstudio/DSView/DSView/pv/mainwindow.cpp: In member function ‘bool pv::MainWindow
/home/anatol/sources/dreamstudio/DSView/DSView/pv/mainwindow.cpp:742:77: error: conversion from ‘uint64_t
             s_obj["vdiv"] = QJsonValue::fromVariant(dsoSig->get_vDialValue());
                                                                             ^

+++ DSView/pv/mainwindow.cpp
@@ -738,8 +738,8 @@ bool MainWindow::store_session(QString name)
         s_obj["strigger"] = s->get_trig();
         boost::shared_ptr<view::DsoSignal> dsoSig;
         if (dsoSig = dynamic_pointer_cast<view::DsoSignal>(s)) {
-            s_obj["vdiv"] = QJsonValue::fromVariant(dsoSig->get_vDialValue());
-            s_obj["vfactor"] = QJsonValue::fromVariant(dsoSig->get_factor());
+            s_obj["vdiv"] = QJsonValue::fromVariant(static_cast<qulonglong>(dsoSig->get_vDialValue()));
+            s_obj["vfactor"] = QJsonValue::fromVariant(static_cast<qulonglong>(dsoSig->get_factor()));
             s_obj["coupling"] = dsoSig->get_acCoupling();
             s_obj["trigValue"] = dsoSig->get_trigRate();
             s_obj["zeroPos"] = dsoSig->get_zeroRate();
